+++ src/sites/link/sylnk.net.js
@@ -40,7 +40,8 @@ _.register({
         /^gadget13\.com$/,
         /^azhie\.net$/,
         /^ww2\.awsubs\.co$/,
-        /^autorp\.us$/
+        /^autorp\.us$/,
+        /^plantaheim\.web\.id$/,
       ],
       query: /^\?d=([a-zA-Z0-9/=]+)$/,
     },
@@ -83,6 +84,10 @@ _.register({
       path: /^\/go2/,
       query: /^\?p=([a-zA-Z0-9/=]+)$/,
     },
+    {
+      host: /^animeforce\.stream$/,
+      query: /^\?l=([a-zA-Z0-9/=]+)$/,
+    },
   ],
   async start (m) {
     const rawLink = atob(m.query[1]);
@@ -104,7 +109,6 @@ _.register({
         // safelinkreviewer.com
         // safelinkreview.co
         /^safelink(s?review(er)?)\.com?$/,
-        /^susutin\.com$/,
         /^(getcomics|miuitutorial)\.gq$/,
         /^awsubs\.cf$/,
         /^awsubsco\.ga$/,
@@ -118,7 +122,7 @@ _.register({
         /^ww3\.manteb\.in$/,
         /^link\.filmku\.net$/,
         /^www\.muucih\.com$/,
-        /^(naisho|filmku)\.lompat\.in$/,
+        /^(naisho|filmku|henpoi)\.lompat\.in$/,
         /^edogawa\.lon\.pw$/,
         /^telolet\.in$/,
       ],
@@ -158,8 +162,10 @@ _.register({
       host: [
         /^(designinghomey|ani-share|sinopsisfilmku|autolinkach)\.com$/,
         /^motonews\.club$/,
-        /^(autofans|landscapenature)\.pw$/,
+        /^(autofans|landscapenature|apasih|gameinfo)\.pw$/,
         /^(sidespace|erogedownload)\.net$/,
+        /^otoviral\.racing$/,
+        /^www\.lifesurance\.info$/,
       ],
       query: /get=([^&]+)/,
     },
@@ -202,6 +208,16 @@ _.register({
     const a = $('div.kiri > center > a');
     await $.openLink(a.href);
   },
+});
+
+_.register({
+  rule: {
+    host: /^aw-games\.net$/,
+  },
+  async ready () {
+    const a = $('.iklan a');
+    await $.openLink(a.href);
+  }
 });
 
 _.register({